More client refactoring. Verified checkmark now appears without having to refresh the page.
