+++ api/controllers/HomeController.js
@@ -12,7 +12,7 @@ module.exports = {
     res.view();
     Reddit.getBothFlairs(sails.config.reddit.adminRefreshToken, req.user.name, function (err, flair1, flair2) {
       if ((flair1 || flair2)) {
-        user.flair = {ptrades: flair1, svex: flair2};
+        req.user.flair = {ptrades: flair1, svex: flair2};
         var ptrades_fcs, svex_fcs;
         if (flair1 && flair1.flair_text) {
           ptrades_fcs = flair1.flair_text.match(/(\d{4}-){2}\d{4}/g);
@@ -20,8 +20,8 @@ module.exports = {
         if (flair2 && flair2.flair_text) {
           svex_fcs = flair2.flair_text.match(/(\d{4}-){2}\d{4}/g);
         }
-        user.loggedFriendCodes = _.union(ptrades_fcs, svex_fcs, user.loggedFriendCodes);
-        user.save(function (err) {
+        req.user.loggedFriendCodes = _.union(ptrades_fcs, svex_fcs, req.user.loggedFriendCodes);
+        req.user.save(function (err) {
           if (err) {
             console.log(err);
           }
