+++ views/home/banlist.ejs
@@ -12,7 +12,7 @@
           </div>
           <div class="col-md-3">
             <button type="button" class="btn btn-danger"
-                    ng-click="setLocalBan(user, false)">Unban User</button>
+                    ng-click="setLocalBan(user.name, false)">Unban User</button>
           </div>
         </li>
       </ul>
