+++ views/home/header.ejs
@@ -72,7 +72,7 @@
         <li class="dropdown">
           <a href="#" class="dropdown-toggle" data-toggle="dropdown">
             <span ng-class="'flair-' + user.flair.ptrades.flair_css_class"></span>
-            <span ng-class="getFlairTextForUserForSVEx()">
+            <span ng-class="getFlairTextForSVEx()">
                 {{user.name}}
             </span>
             <span class="caret"></span>
