+++ views/home/references.ejs
@@ -47,7 +47,7 @@
           <span class="glyphicon glyphicon-edit"></span>
         </a>
         <a href="#" ng-really-switch="deleteRef" class="icon"
-           ng-really-click="deleteRef(reference.id, reference, 'events')"
+           ng-really-click="deleteRef(reference.id)"
            ng-if="(!refUser || (user.name === refUser.name) || user.isMod)">
           <span class="glyphicon glyphicon-trash"></span>
         </a>
@@ -96,7 +96,7 @@
           <span class="glyphicon glyphicon-edit"></span>
         </a>
         <a href="#"  ng-really-switch="deleteRef" class="icon"
-           ng-really-click="deleteRef(reference.id, reference, 'shinies')"
+           ng-really-click="deleteRef(reference.id)"
            ng-if="(!refUser || (user.name === refUser.name) || user.isMod)">
           <span class="glyphicon glyphicon-trash"></span>
         </a>
@@ -145,7 +145,7 @@
           <span class="glyphicon glyphicon-edit"></span>
         </a>
         <a href="#"  ng-really-switch="deleteRef" class="icon"
-           ng-really-click="deleteRef(reference.id, reference, 'casuals')"
+           ng-really-click="deleteRef(reference.id)"
            ng-if="(!refUser || (user.name === refUser.name) || user.isMod)">
           <span class="glyphicon glyphicon-trash"></span>
         </a>
@@ -177,7 +177,7 @@
           <span class="glyphicon glyphicon-edit"></span>
         </a>
         <a href="#"  ng-really-switch="deleteRef" class="icon"
-           ng-really-click="deleteRef(reference.id, reference, 'banks')"
+           ng-really-click="deleteRef(reference.id)"
            ng-if="(!refUser || (user.name === refUser.name) || user.isMod)">
           <span class="glyphicon glyphicon-trash"></span>
         </a>
@@ -225,7 +225,7 @@
           <span class="glyphicon glyphicon-edit"></span>
         </a>
         <a href="#"  ng-really-switch="deleteRef" class="icon"
-           ng-really-click="deleteRef(reference.id, reference, 'eggs')"
+           ng-really-click="deleteRef(reference.id)"
            ng-if="(!refUser || (user.name === refUser.name) || user.isMod)">
           <span class="glyphicon glyphicon-trash"></span>
         </a>
@@ -275,7 +275,7 @@
           <span class="glyphicon glyphicon-edit"></span>
         </a>
         <a href="#"  ng-really-switch="deleteRef" class="icon"
-           ng-really-click="deleteRef(reference.id, reference, 'giveaways')"
+           ng-really-click="deleteRef(reference.id)"
            ng-if="(!refUser || (user.name === refUser.name) || user.isMod)">
           <span class="glyphicon glyphicon-trash"></span>
         </a>
@@ -327,7 +327,7 @@
                   <span class="glyphicon glyphicon-edit"></span>
               </a>
               <a href="#"  ng-really-switch="deleteRef" class="icon"
-                 ng-really-click="deleteRef(reference.id, reference, 'eggchecks')"
+                 ng-really-click="deleteRef(reference.id)"
                  ng-if="(!refUser || (user.name === refUser.name) || user.isMod)">
                   <span class="glyphicon glyphicon-trash"></span>
               </a>
@@ -379,7 +379,7 @@
                   <span class="glyphicon glyphicon-edit"></span>
               </a>
               <a href="#"  ng-really-switch="deleteRef" class="icon"
-                 ng-really-click="deleteRef(reference.id, reference, 'eggchecks')"
+                 ng-really-click="deleteRef(reference.id)"
                  ng-if="(!refUser || (user.name === refUser.name) || user.isMod)">
                   <span class="glyphicon glyphicon-trash"></span>
               </a>
@@ -418,7 +418,7 @@
           <span class="glyphicon glyphicon-edit"></span>
         </a>
         <a href="#"  ng-really-switch="deleteRef" class="icon"
-           ng-really-click="deleteRef(reference.id, reference, 'misc')"
+           ng-really-click="deleteRef(reference.id)"
            ng-if="(!refUser || (user.name === refUser.name) || user.isMod)">
           <span class="glyphicon glyphicon-trash"></span>
         </a>
